convert header
